move class for test data
